Adds demo option for handling json
Since we're switching to handling straight json instead of jsonp.
this PR needs to be coordinated with the `format-json` branch
of CCDB4/complaint-randomizer.

Originally the page JS pulled the json file straight from S3, which
required formatting the narratives as jsonp to avoid CORS problems.
We can drop all that now.