Adds demo option for handling json
Since we're switching to handling straight json instead of jsonp,
this PR needs to be coordinated with the `json-handling` branch
of CCDB4/complaint-randomizer.

The original WordPress page pulled the json straight from S3, which
required formatting the narratives as jsonp to avoid CORS restrictions.
We can drop all that now, because a Django view pulls from S3.