remove deprecated narrative formatting
The current logic for the complaint database landing page invokes a
format_narratives function that does some processing on narrative data
pulled down from S3.

That data currently causes the following error to be logged on every
single load of the landing page:

    format_narratives:KeyError
    There is problem accessing with the given key, which probably means the json has missing data
    'date_received'

This error gets printed because each narrative in the current JSON data
doesn't necessarily contain that key.

In practice, this doesn't matter at all because this narrative data is
never used on the landing page. It gets processed into a "narrative"
variable that gets passed into the template, but that variable isn't
used at all.

The JSON data is still used for stats to present the number of
complaints that were resolved in a timely manner, so we still need to
keep fetching of the JSON. But if the narrative data itself isn't
displayed at all, can this logic (and the frequently-appearing error) be
removed?

One other minor change: the formatting of .travis.yml has been changed
so that multiple environment variables are used globally, instead of
causing two separate Travis jobs.